User Data connection to front end
The user data will now be displayed on the front end of the website, due to changes made in the related .ejs files to display the data.
In addition the api GET Request was updated and allows you to get a specific user if the there is an id avialable.